requirements.txt
pandas>=2.2.3
requests>=2.32.3
rasterio>=1.4.3
numpy>=2.2.3
tqdm>=4.67.1
Pillow>=11.1.0
